Initial types extraction
